make Test Outcomes inherit from BaseException instead of exception
fixes #580
