+++ _pytest/recwarn.py
@@ -7,8 +7,9 @@ import _pytest._code
 import py
 import sys
 import warnings
-from _pytest.fixtures import yield_fixture
 
+from _pytest.fixtures import yield_fixture
+from _pytest.outcomes import fail
 
 @yield_fixture
 def recwarn():
@@ -197,7 +198,6 @@ class WarningsChecker(WarningsRecorder):
                 if not any(issubclass(r.category, self.expected_warning)
                            for r in self):
                     __tracebackhide__ = True
-                    from _pytest.runner import fail
                     fail("DID NOT WARN. No warnings of type {0} was emitted. "
                          "The list of emitted warnings is: {1}.".format(
                              self.expected_warning,
